Factor Examples.swift + GraphType.swift out of ViewController.swift
